+++ tests/SlmQueueTest/Job/JobTest.php
@@ -15,41 +15,6 @@ class JobTest extends TestCase
         $this->assertEquals(3, $job->getMetadata('id'));
     }
 
-    public function testCorrectlySerializeJobContent()
-    {
-        $job = new SimpleJob();
-        $job->setContent('Foo');
-
-        $this->assertEquals('{"class":"SlmQueueTest\\\Asset\\\SimpleJob","content":"s:3:\"Foo\";","metadata":[]}', $job->jsonSerialize());
-    }
-
-    public function testCorrectlySerializeJobMetadata()
-    {
-        $job = new SimpleJob();
-        $job->setMetadata('foo', 'Bar');
-
-        $this->assertEquals('{"class":"SlmQueueTest\\\Asset\\\SimpleJob","content":"N;","metadata":{"foo":"Bar"}}', $job->jsonSerialize());
-    }
-
-    public function testCorrectlySerializeJobContentAndMetadata()
-    {
-        $job = new SimpleJob();
-        $job->setContent('Foo');
-        $job->setMetadata('foo', 'Bar');
-
-        $this->assertEquals('{"class":"SlmQueueTest\\\Asset\\\SimpleJob","content":"s:3:\"Foo\";","metadata":{"foo":"Bar"}}', $job->jsonSerialize());
-    }
-
-    public function testCorrectlyUnserializeJob()
-    {
-        $job = new SimpleJob();
-        $job->setContent('Foo');
-        $job = json_decode($job->jsonSerialize(), true);
-
-        $this->assertEquals('SlmQueueTest\Asset\SimpleJob', $job['class']);
-        $this->assertEquals('Foo', unserialize($job['content']));
-    }
-
     public function testJobCanBeExecuted()
     {
         // The simple Job just add a metadata