+++ tests/SlmQueueTest/Queue/QueueTest.php
@@ -10,98 +10,180 @@ use SlmQueueTest\Asset\SimpleJob;
 
 class QueueTest extends TestCase
 {
+    protected $job;
+    protected $jobName;
+    protected $jobPluginManager;
+    protected $queue;
+
+    public function setUp()
+    {
+        $this->job     = new SimpleJob;
+        $this->jobName = 'SlmQueueTest\Asset\SimpleJob';
+
+        $this->jobPluginManager = $this->getMock('SlmQueue\Job\JobPluginManager');
+        $this->queue = new SimpleQueue('queue', $this->jobPluginManager);
+    }
+
     public function testCanPushThenPopJob()
     {
-        $jobPluginManager = $this->getMock('SlmQueue\Job\JobPluginManager');
-        $jobPluginManager->expects($this->once())
-                         ->method('get')
-                         ->with('SlmQueueTest\Asset\SimpleJob')
-                         ->will($this->returnValue(new SimpleJob()));
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with($this->jobName)
+                               ->will($this->returnValue($this->job));
 
-        $queue = new SimpleQueue('queue', $jobPluginManager);
+        $this->queue->push($this->job);
+        $job = $this->queue->pop();
 
+        $this->assertInstanceOf($this->jobName, $job);
+
+        $expected = spl_object_hash($this->job);
+        $actual   = spl_object_hash($job);
+        $this->assertEquals($expected, $actual);
+    }
+
+    public function testCanPushThenPopWithJobContent()
+    {
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with($this->jobName)
+                               ->will($this->returnValue($this->job));
+
+        $this->job->setContent('Foo');
+
+        $this->queue->push($this->job);
+        $job = $this->queue->pop();
+
+        $this->assertEquals('Foo', $job->getContent());
+    }
+
+    public function testCanPushThenPopWithJobMetadata()
+    {
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with($this->jobName)
+                               ->will($this->returnValue($this->job));
+
+        $this->job->setMetadata('Foo', 'Bar');
+
+        $this->queue->push($this->job);
+        $job = $this->queue->pop();
+
+        $this->assertEquals(array('Foo' => 'Bar'), $job->getMetadata());
+        $this->assertEquals('Bar', $job->getMetadata('Foo'));
+    }
+
+    public function testCorrectlySerializeJobContent()
+    {
         $job = new SimpleJob();
         $job->setContent('Foo');
 
-        $queue->push($job);
+        $expected = '{"name":"SlmQueueTest\\\Asset\\\SimpleJob","content":"s:3:\"Foo\";","metadata":[]}';
+        $actual   = $this->queue->serializeJob($job);
 
-        $job = $queue->pop();
+        $this->assertEquals($expected, $actual);
+    }
 
-        $this->assertInstanceOf('SlmQueueTest\Asset\SimpleJob', $job);
-        $this->assertEquals('Foo', $job->getContent());
+    public function testCorrectlySerializeJobMetadata()
+    {
+        $job = new SimpleJob();
+        $job->setMetadata('Foo', 'Bar');
+
+        $expected = '{"name":"SlmQueueTest\\\Asset\\\SimpleJob","content":"N;","metadata":{"Foo":"Bar"}}';
+        $actual   = $this->queue->serializeJob($job);
+
+        $this->assertEquals($expected, $actual);
+    }
+
+    public function testCorrectlySerializeJobContentAndMetadata()
+    {
+        $job = new SimpleJob();
+        $job->setContent('Foo');
+        $job->setMetadata('Foo', 'Bar');
+
+        $expected = '{"name":"SlmQueueTest\\\Asset\\\SimpleJob","content":"s:3:\"Foo\";","metadata":{"Foo":"Bar"}}';
+        $actual   = $this->queue->serializeJob($job);
+
+        $this->assertEquals($expected, $actual);
+    }
+
+    public function testCorrectlySerializeJobServiceName()
+    {
+        $job = new SimpleJob();
+        $job->setMetadata('name', 'SimpleJob');
+
+        $expected = '{"name":"SimpleJob","content":"N;","metadata":{"name":"SimpleJob"}}';
+        $actual   = $this->queue->serializeJob($job);
+
+        $this->assertEquals($expected, $actual);
+    }
+
+    public function testCanCreateJobWithFQCN()
+    {
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with($this->jobName)
+                               ->will($this->returnValue($this->job));
+
+        $payload = '{"name":"SlmQueueTest\\\Asset\\\SimpleJob","content":"N;","metadata":[]}';
+        $job     = $this->queue->unserializeJob($payload);
+
+        $expected = spl_object_hash($this->job);
+        $actual   = spl_object_hash($job);
+        $this->assertEquals($expected, $actual);
     }
 
     public function testCanCreateJobWithStringName()
     {
-        $job = new SimpleJob();
-        $jobPluginManager = $this->getMock('SlmQueue\Job\JobPluginManager');
-        $jobPluginManager->expects($this->once())
-                         ->method('get')
-                         ->with('SimpleJob')
-                         ->will($this->returnValue($job));
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with('SimpleJob')
+                               ->will($this->returnValue($this->job));
 
-        $queue  = new SimpleQueue('queue', $jobPluginManager);
-        $result = $queue->createJob('SimpleJob');
+        $payload = '{"name":"SimpleJob","content":"N;","metadata":[]}';
+        $job     = $this->queue->unserializeJob($payload);
 
-        $expected = spl_object_hash($job);
-        $actual   = spl_object_hash($result);
+        $expected = spl_object_hash($this->job);
+        $actual   = spl_object_hash($job);
         $this->assertEquals($expected, $actual);
     }
 
     public function testCanCreateJobWithContent()
     {
-        $job = $this->getMock('SlmQueue\Job\JobInterface');
-        $job->expects($this->once())
-            ->method('setContent')
-            ->with('Foo')
-            ->will($this->returnValue($job));
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with($this->jobName)
+                               ->will($this->returnValue($this->job));
 
-        $jobPluginManager = $this->getMock('SlmQueue\Job\JobPluginManager');
-        $jobPluginManager->expects($this->once())
-                         ->method('get')
-                         ->with('SimpleJob')
-                         ->will($this->returnValue($job));
+        $payload = '{"name":"SlmQueueTest\\\Asset\\\SimpleJob","content":"s:3:\"Foo\";","metadata":[]}';
+        $job     = $this->queue->unserializeJob($payload);
 
-        $queue  = new SimpleQueue('queue', $jobPluginManager);
-        $result = $queue->createJob('SimpleJob', serialize('Foo'));
+        $this->assertEquals('Foo', $job->getContent());
     }
 
     public function testCanCreateJobWithMetadata()
     {
-        $job = $this->getMock('SlmQueue\Job\JobInterface');
-        $job->expects($this->once())
-            ->method('setContent')
-            ->will($this->returnValue($job));
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with($this->jobName)
+                               ->will($this->returnValue($this->job));
 
-        $job->expects($this->once())
-            ->method('setMetadata')
-            ->with(array('foo' => 'Bar'))
-            ->will($this->returnValue($job));
+        $payload = '{"name":"SlmQueueTest\\\Asset\\\SimpleJob","content":"N;","metadata":{"Foo":"Bar"}}';
+        $job     = $this->queue->unserializeJob($payload);
 
-        $jobPluginManager = $this->getMock('SlmQueue\Job\JobPluginManager');
-        $jobPluginManager->expects($this->once())
-                         ->method('get')
-                         ->with('SimpleJob')
-                         ->will($this->returnValue($job));
-
-        $queue  = new SimpleQueue('queue', $jobPluginManager);
-        $result = $queue->createJob('SimpleJob', null, array('foo' => 'Bar'));
+        $this->assertEquals('Bar', $job->getMetadata('Foo'));
     }
 
     public function testCreateQueueAwareJob()
     {
         $job = new QueueAwareJob();
+        $this->jobPluginManager->expects($this->once())
+                               ->method('get')
+                               ->with('QueueAwareJob')
+                               ->will($this->returnValue($job));
 
-        $jobPluginManager = $this->getMock('SlmQueue\Job\JobPluginManager');
-        $jobPluginManager->expects($this->once())
-                         ->method('get')
-                         ->with('SlmQueueTest\Asset\QueueAwareJob')
-                         ->will($this->returnValue($job));
+        $payload = '{"name":"QueueAwareJob","content":"N;","metadata":{"name":"QueueAwareJob"}}';
+        $result  = $this->queue->unserializeJob($payload);
 
-        $queue = new SimpleQueue('queue', $jobPluginManager);
-
-        $result = $queue->createJob('SlmQueueTest\Asset\QueueAwareJob');
-
-        $this->assertSame($queue, $result->getQueue());
+        $this->assertSame($this->queue, $job->getQueue());
     }
 }